kitakami-common: workaround: disable graphics driver preloading
As the Mali userspace binaries were built for Android 8.0,
when -ffixed-x18 was not guarteed to be used, which had been
ensured for Android versions newer than P for the shadow
call stack (SCS) feature.
For such old version vendor blob, there was a workaround
in com_android_internal_os_ZygoteInit.cpp which has been removed
now.
    https://android-review.googlesource.com/c/platform/frameworks/base/+/2518435

After that removal, there is a crash on libbluetooth_jni.so
reported as the issue here:
    https://issuetracker.google.com/issues/302408537

As a workaround, disabling the preloading of them to avoid
the bluetooth crash issue.

Bug: 302408537
Test: bluetooth enabled by default after booted up

Change-Id: Icd8be5e120b016b3019e0d0dd23121f7e931a67c
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>

@acroreiser: the same but with Adreno binaries.